+++ src/main/kotlin/com/knu/mockin/service/BasicRealService.kt
@@ -3,9 +3,12 @@ package com.knu.mockin.service
 import com.knu.mockin.exeption.ErrorCode
 import com.knu.mockin.kisclient.KISBasicRealClient
 import com.knu.mockin.model.dto.kisheader.request.KISOverSeaRequestHeaderDto
-import com.knu.mockin.model.dto.kisrequest.basic.*
-import com.knu.mockin.model.dto.kisresponse.basic.*
-import com.knu.mockin.model.dto.request.basic.*
+import com.knu.mockin.model.dto.kisrequest.quotations.analysis.KISNewsTitleRequestParameterDto
+import com.knu.mockin.model.dto.kisrequest.quotations.basic.real.*
+import com.knu.mockin.model.dto.kisresponse.quotations.analysis.KISNewsTitleResponseDto
+import com.knu.mockin.model.dto.kisresponse.quotations.basic.real.*
+import com.knu.mockin.model.dto.request.quotations.analysis.real.NewsTitleRequestParameterDto
+import com.knu.mockin.model.dto.request.quotations.basic.real.*
 import com.knu.mockin.model.enum.Constant.REAL
 import com.knu.mockin.model.enum.TradeId
 import com.knu.mockin.repository.RealKeyRepository
